Projeto Final

+++ Projeto_Final/TropPizza/server/TropPizza.Domain/Features/Customers/ICustomerRepository.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 
 namespace TropPizza.Domain.Features.Customers
@@ -5,10 +6,10 @@ namespace TropPizza.Domain.Features.Customers
     public interface ICustomerRepository
     {
         public void Create(Customer customer);
-        public Customer ReadById(int id);
+        public Customer ReadById(Int64 id);
         // public Customer ReadByCpf(string cpf);
         public List<Customer> ReadAll();
         public void Update(Customer customer);
-        public void Delete(int id);
+        public void Delete(Int64 id);
     }
 }
